update: Working L1 & Gateway resilience
