+++ madara/crates/client/settlement_client/Cargo.toml
@@ -27,6 +27,7 @@ mp-block.workspace = true
 mp-chain-config.workspace = true
 mp-convert.workspace = true
 mp-oracle.workspace = true
+mp-resilience.workspace = true
 mp-transactions.workspace = true
 mp-utils.workspace = true
