+++ madara/node/Cargo.toml
@@ -36,6 +36,7 @@ mc-telemetry = { workspace = true }
 mp-block = { workspace = true }
 mp-chain-config = { workspace = true }
 mp-oracle = { workspace = true }
+mp-resilience = { workspace = true }
 mp-rpc = { workspace = true }
 mp-transactions = { workspace = true }
 mp-utils = { workspace = true }
